Fixed bug where powershell window would flash while executing commands on Windows, fixed skelctl bug
For windows bug fix, I needed to use syscall to hide the window. However, the code does not compile for other OSes because it is a windows only feature so I seperated the go files into main.go for all OSes but windows and windows uses main.go. As a result I was also able to cut out the GOOS check when handling tasks to check the OS to see to use powershell. I also updated the Makefile accordingly

For skelctl bug fix, it would create an extra task with a blank agent id when running "skelctl cmd <command>". It now does not create a task if agent_id is blank